Seperated beatmap- and skin datamodels from structures- and datacontainers holding the raw data
